v3

+++ MX/mx_donuts/donuts.py
@@ -85,7 +85,7 @@ def find_flour_needed_for_baking(donut_list: list[Donut], quantity: int) -> int:
     :param quantity: The quantity of each donut to be baked.
     :return: Total amount of flour needed in grams rounded up.
     """
-    return sum(map(lambda x: ((x.price * 75) * quantity), donut_list))
+    return sum(map(lambda x: int(((x.price * 75) * quantity)), donut_list))
 
 
 def calculate_tip(donut_list: list[Donut], customers: int) -> int:
@@ -98,7 +98,7 @@ def calculate_tip(donut_list: list[Donut], customers: int) -> int:
     :param customers: Number of customers visiting bakery.
     :return: Tip amount.
     """
-    return sum(map(lambda x: (x.price * 0.2 * customers), donut_list))
+    return sum(map(lambda x: int((x.price * 0.2 * customers)), donut_list))
 
 
 def sort_donuts_by_allergies(donut_list: list[Donut]) -> list[str]:
